simplify mypylib
drop support of config.memoryUsinglimit
