+++ mypylib/logger.py
@@ -19,6 +19,7 @@ _MODE_TO_LEVEL = {
     ERROR: logging.ERROR,
 }
 
+
 def level_for_mode(mode: str) -> int:
     return _MODE_TO_LEVEL.get(mode, logging.INFO)
 
@@ -41,20 +42,26 @@ class LogFormatter(logging.Formatter):
 
     def format(self, record: logging.LogRecord) -> str:
         mode = record.levelname.lower()
-        dt = date_time_library.datetime.fromtimestamp(record.created, date_time_library.timezone.utc)
+        dt = date_time_library.datetime.fromtimestamp(
+            record.created, date_time_library.timezone.utc
+        )
         time_text = dt.strftime("%d.%m.%Y, %H:%M:%S.%f")[:-3]
-        time_text = "{0} (UTC)".format(time_text).ljust(32, ' ')
-        mode_field = "[{0}]".format(mode).ljust(10, ' ')
-        thread_field = "<{0}>".format(record.threadName).ljust(14, ' ')
+        time_text = f"{time_text} (UTC)".ljust(32, " ")
+        mode_field = f"[{mode}]".ljust(10, " ")
+        thread_field = f"<{record.threadName}>".ljust(14, " ")
         message = record.getMessage()
         if record.exc_info:
-            message = message + '\n' + self.formatException(record.exc_info)
+            message = message + "\n" + self.formatException(record.exc_info)
         if not self.colored:
             return mode_field + time_text + thread_field + message
-        mode_color = self._LEVEL_COLORS.get(record.levelno, bcolors.UNDERLINE) + bcolors.BOLD
-        mode_text = "{0}{1}{2}".format(mode_color, mode_field, bcolors.ENDC)
-        thread_color = (bcolors.ERROR if record.levelno >= logging.ERROR else bcolors.OKGREEN) + bcolors.BOLD
-        thread_text = "{0}{1}{2}".format(thread_color, thread_field, bcolors.ENDC)
+        mode_color = (
+            self._LEVEL_COLORS.get(record.levelno, bcolors.UNDERLINE) + bcolors.BOLD
+        )
+        mode_text = f"{mode_color}{mode_field}{bcolors.ENDC}"
+        thread_color = (
+            bcolors.ERROR if record.levelno >= logging.ERROR else bcolors.OKGREEN
+        ) + bcolors.BOLD
+        thread_text = f"{thread_color}{thread_field}{bcolors.ENDC}"
         return mode_text + time_text + thread_text + message
 
 
@@ -64,7 +71,11 @@ def get_logger(name: str | None = None) -> logging.Logger:
     return logging.getLogger(f"{ROOT_LOGGER_NAME}.{name}")
 
 
-def setup_logging(log_level: str = INFO, log_file_name: str | None = None, log_limit_lines: int | None = None) -> logging.Logger:
+def setup_logging(
+    log_level: str = INFO,
+    log_file_name: str | None = None,
+    log_limit_lines: int | None = None,
+) -> logging.Logger:
     logger = logging.getLogger(ROOT_LOGGER_NAME)
     logger.setLevel(level_for_mode(log_level))  # child loggers inherit this level
     for handler in list(logger.handlers):
